# internal/delivery/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/specialist/create_rated": {
            "post": {
                "description": "Creates a new rating entry based on the provided data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specialists"
                ],
                "parameters": [
                    {
                        "description": "Rated data",
                        "name": "rated_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RatedCreate"
                        }
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created the rating",
                        "schema": {
                            "$ref": "#/definitions/responses.CreationIntResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/specialist/get_cases_by_level": {
            "get": {
                "description": "Retrieves cases based on the provided cursor ID and the user's ID. It returns cases that match the level of difficulty or rating specified for the user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specialists"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cursor ID for pagination",
                        "name": "cursor_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the cases by level",
                        "schema": {
                            "$ref": "#/definitions/models.CaseCursor"
                        }
                    },
                    "400": {
                        "description": "Invalid input data or bad query parameter",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "403": {
                        "description": "User is unverified",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/specialist/get_rated_solved": {
            "get": {
                "description": "Retrieves a rated solved entry based on the provided cursor ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specialists"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cursor ID for pagination",
                        "name": "cursor_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the rated solved",
                        "schema": {
                            "$ref": "#/definitions/models.RatedCursor"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/specialist/me": {
            "get": {
                "description": "Retrieves information about the current specialist based on their user ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specialists"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved the specialist info",
                        "schema": {
                            "$ref": "#/definitions/models.Specialist"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        },
        "/specialist/update": {
            "put": {
                "description": "Updates an existing specialist's information including their password, full name, and photo.\nThe password must be more than 8 symbols and contain at least one number, one uppercase, and one lowercase letter.\nThe photo upload is optional but must be a valid image file if provided.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "specialists"
                ],
                "summary": "UpdateMain Specialist Information with Photo Upload",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Full Name",
                        "name": "fullname",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Photo Upload",
                        "name": "photo",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful update, no content returned"
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "JWT is invalid or expired",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error, could not process the request",
                        "schema": {
                            "$ref": "#/definitions/responses.MessageResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Case": {
            "type": "object",
            "properties": {
                "camera_id": {
                    "type": "string"
                },
                "datetime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "photo_url": {
                    "type": "string"
                },
                "transport": {
                    "type": "string"
                },
                "violation_id": {
                    "type": "string"
                },
                "violation_value": {
                    "type": "string"
                }
            }
        },
        "models.CaseCursor": {
            "type": "object",
            "properties": {
                "cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Case"
                    }
                },
                "cursor": {
                    "$ref": "#/definitions/null.Int"
                }
            }
        },
...
        "models.Rated": {
            "type": "object",
            "required": [
                "case_id"
            ],
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "choice": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "specialist_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.RatedCreate": {
            "type": "object",
            "required": [
                "case_id"
            ],
            "properties": {
                "case_id": {
                    "type": "integer"
                },
                "choice": {
                    "type": "boolean"
                }
            }
        },
        "models.RatedCursor": {
            "type": "object",
            "properties": {
                "cursor": {
                    "$ref": "#/definitions/null.Int"
                },
                "rated": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Rated"
                    }
                }
            }
        },
...
        "models.Specialist": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "fullname": {
                    "$ref": "#/definitions/null.String"
                },
                "id": {
                    "type": "integer"
                },
                "isVerified": {
                    "type": "boolean"
                },
                "level": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photoUrl": {
                    "$ref": "#/definitions/null.String"
                }
            }
        },
...
        "null.Int": {
            "type": "object",
            "properties": {
                "int64": {
                    "type": "integer"
                },
                "valid": {
                    "description": "Valid is true if Int64 is not NULL",
                    "type": "boolean"
                }
            }
        },
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "responses.CreationIntResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "responses.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}